+++ itables/sample_dfs.py
@@ -375,3 +375,38 @@ def generate_random_df(rows, columns, column_types=COLUMN_TYPES):
         x.index = index
 
     return pd.DataFrame(series)
+
+
+def get_pandas_styler():
+    """This function returns a Pandas Styler object
+
+    Cf. https://pandas.pydata.org/docs/user_guide/style.html
+    """
+    x = np.linspace(0, math.pi, 21)
+    df = pd.DataFrame(
+        {"sin": np.sin(x), "cos": np.cos(x)}, index=pd.Index(x, name="alpha")
+    )
+
+    s = df.style
+    s.background_gradient(axis=None, cmap="YlOrRd")
+    s.format("{:.3f}")
+
+    s.set_caption(
+        "A Pandas Styler object with background colors and tooltips"
+    ).set_table_styles(
+        [{"selector": "caption", "props": "caption-side: bottom; font-size:1.25em;"}],
+    )
+
+    ttips = pd.DataFrame(
+        {
+            "sin": ["The sinus of {:.6f} is {:.6f}".format(t, np.sin(t)) for t in x],
+            "cos": ["The cosinus of {:.6f} is {:.6f}".format(t, np.cos(t)) for t in x],
+        },
+        index=df.index,
+    )
+    try:
+        s.set_tooltips(ttips)
+    except AttributeError:
+        pass
+
+    return s